update to win 4.2.0/mac 1.2.0 version

+++ src/prism-live-studio/PRISMLiveStudio/prism-ui/channels/ChannelsDataApi/PLSChannelDataAPI.cpp
@@ -108,12 +108,14 @@ int PLSChannelDataAPI::countTask(const QString &taskName) const
 
 void PLSChannelDataAPI::endTransactions()
 {
+	PLS_INFO("Channel", "end transactions");
 	if (mTransactions.isEmpty()) {
+		updatePlatformViewerCount();
 		return;
 	}
 
-	PLSCHANNELS_API->sigAllChannelRefreshDone();
-
+	PLSCHANNELS_API->sigOperationChannelDone();
+	PLS_INFO("Channel", "Operation Channel Done");
 	QVariantList tasks = mTransactions.value(ChannelTransactionsKeys::g_taskQueue).toList();
 	if (tasks.isEmpty()) {
 		mTransactions.clear();
@@ -191,6 +193,12 @@ void PLSChannelDataAPI::connectSignals()
 	auto handlePrismExpired = []() { reloginPrismExpired(); };
 	connect(this, &PLSChannelDataAPI::prismTokenExpired, qApp, handlePrismExpired, Qt::QueuedConnection);
 
+	auto handleB2BExpired = [](const QString &channelName) {
+		PLS_INFO("Channel", "b2b token expired when in live, end living will logout");
+		pls_alert_error_message(getMainWindow(), QObject::tr("Alert.Title"), QObject::tr("Ncb2b.Token.Expired").arg(channelName));
+		pls_prism_change_over_login_view();
+	};
+	connect(this, &PLSChannelDataAPI::sigB2BChannelEndLivingCheckExpired, qApp, handleB2BExpired, Qt::QueuedConnection);
 	//in shifting
 	connect(PLS_PLATFORM_API, &PLSPlatformApi::livePrepared, this, [this](bool isOk) {
 		if (isOk) {
@@ -326,10 +334,19 @@ QPixmap PLSChannelDataAPI::updateImage(const QString &oldPath, const QString &sr
 QPixmap PLSChannelDataAPI::updateImage(const QString &srcPath, const QSize &size)
 {
 	QFileInfo fInfo(srcPath);
+
+	//don't remove this folder
+	bool bRemove = true;
+	QString tmpPath = QString("PRISMLiveStudio/library/library_Policy_PC/images");
+	tmpPath = pls_get_user_path(tmpPath);
+	if (srcPath.contains(tmpPath, Qt::CaseInsensitive)) {
+		bRemove = false;
+	}
+
 	QVariantMap imageMap = mImagesCache.value(fInfo.fileName());
 	QPixmap tmp;
 	if (!imageMap.isEmpty()) {
-		if (fInfo.isWritable()) {
+		if (fInfo.isWritable() && bRemove) {
 			QFile::remove(srcPath);
 		}
 		// to get target size image
@@ -339,7 +356,7 @@ QPixmap PLSChannelDataAPI::updateImage(const QString &srcPath, const QSize &size
 	//try to load source image
 	loadPixmap(tmp, srcPath, size);
 
-	if (fInfo.isWritable()) {
+	if (fInfo.isWritable() && bRemove) {
 		QFile::remove(srcPath);
 	}
 	//if load fail return null
@@ -407,10 +424,10 @@ void PLSChannelDataAPI::updateChannelInfo(const QString &channelUUID, const QVar
 {
 
 	QWriteLocker locker(&mInfosLocker);
-	auto platName = getInfo(Infos, g_platformName);
+	auto platName = getInfo(Infos, g_channelName);
 	platName = toPlatformCodeID(platName, true);
 	auto tmp = Infos;
-	tmp[g_platformName] = platName;
+	tmp[g_channelName] = platName;
 	mChannelInfos[channelUUID] = tmp;
 	mChannelInfos[channelUUID][g_isPresetRTMP] = PLSSyncServerManager::instance()->isPresetRTMP(getInfo(Infos, g_channelRtmpUrl));
 }
@@ -454,10 +471,11 @@ void PLSChannelDataAPI::reCheckExpiredChannels()
 	auto ite = mExpiredChannels.begin();
 	while (ite != mExpiredChannels.end()) {
 		const auto &info = ite.value();
-		auto platformName = getInfo(info, g_platformName);
+		auto platformName = getInfo(info, g_channelName);
+		auto errorString = getInfo(info, g_errorString);
 		this->removeChannelsByPlatformName(platformName, ChannelDataType::ChannelType, false, false);
 		QMetaObject::invokeMethod(
-			getMainWindow(), [platformName]() { reloginChannel(platformName, true); }, Qt::QueuedConnection);
+			getMainWindow(), [platformName, errorString]() { reloginChannel(platformName, true, errorString); }, Qt::QueuedConnection);
 		++ite;
 	}
 	mExpiredChannels.clear();
@@ -498,8 +516,8 @@ void PLSChannelDataAPI::finishAdding(const QString &channelUUID)
 		exclusiveChannelCheckAndVerify(info);
 		int childrenSelected = 0;
 		if (myType == ChannelType) {
-			childrenSelected = static_cast<int>(this->getCurrentSelectedPlatformChannels(getInfo(info, g_platformName), ChannelType).count());
-			showResolutionTips(getInfo(info, g_platformName));
+			childrenSelected = static_cast<int>(this->getCurrentSelectedPlatformChannels(getInfo(info, g_channelName), ChannelType).count());
+			showResolutionTips(getInfo(info, g_channelName));
 		} else {
 			showResolutionTips(CUSTOM_RTMP);
 		}
@@ -540,8 +558,9 @@ void PLSChannelDataAPI::setChannelInfos(const QVariantMap &Infos, bool notify, b
 	if (notify) {
 		emit channelModified(channelUUID);
 	}
-
-	if (int newUserStatus = getInfo(Infos, g_channelUserStatus, NotExist); getInfo(lastInfo, g_channelUserStatus, NotExist) != newUserStatus) {
+	int newUserStatus = getInfo(Infos, g_channelUserStatus, NotExist);
+	int orgUserStatus = getInfo(lastInfo, g_channelUserStatus, NotExist);
+	if (orgUserStatus != newUserStatus) {
 		emit channelActiveChanged(channelUUID, newUserStatus == Enabled);
 	}
 }
@@ -557,9 +576,9 @@ bool PLSChannelDataAPI::replaceChannelInfo(const QString &channelUUID, QVariantM
 	lastInfo = ite.value();
 	*ite = Infos;
 	(*ite)[g_isPresetRTMP] = PLSSyncServerManager::instance()->isPresetRTMP(getInfo(Infos, g_channelRtmpUrl));
-	auto platName = getInfo(Infos, g_platformName);
+	auto platName = getInfo(Infos, g_channelName);
 	platName = toPlatformCodeID(platName, true);
-	(*ite)[g_platformName] = platName;
+	(*ite)[g_channelName] = platName;
 	return true;
 }
 
@@ -674,7 +693,7 @@ bool PLSChannelDataAPI::isInWholeRecording() const
 	return (currentReocrdState() != RecordReady && currentReocrdState() != RecordStopped);
 }
 
-ChannelsMap PLSChannelDataAPI::getCurrentSelectedChannels(int Type) const
+ChannelsMap PLSChannelDataAPI::getCurrentSelectedChannels(int Type)
 {
 	QVariantMap searchmap;
 	searchmap.insert(g_channelUserStatus, ChannelData::Enabled);
@@ -709,12 +728,12 @@ ChannelsMap PLSChannelDataAPI::getCurrentSelectedPlatformChannels(const QString 
 	auto isMatched = [&](const QVariantMap &info) {
 		int myType = getInfo(info, g_data_type, NoType);
 		bool isTypeMatch = ((srcType == NoType) ? true : (myType == srcType));
-		return getInfo(info, g_platformName).contains(platform) && getInfo(info, g_channelUserStatus, Disabled) == Enabled && isTypeMatch;
+		return getInfo(info, g_channelName).contains(platform) && getInfo(info, g_channelUserStatus, Disabled) == Enabled && isTypeMatch;
 	};
 
 	for (const auto &info : mChannelInfos) {
 		if (isMatched(info)) {
-			retMap.insert(getInfo(info, g_platformName), info);
+			retMap.insert(getInfo(info, g_channelName), info);
 		}
 	}
 	return retMap;
@@ -815,10 +834,11 @@ QVariantMap &PLSChannelDataAPI::getChanelInfoRef(const QString &channelUUID)
 QVariantMap &PLSChannelDataAPI::getChanelInfoRefByPlatformName(const QString &channelName, int type)
 {
 	QReadLocker locker(&mInfosLocker);
+	auto tempName = NCB2BConvertChannelName(channelName);
 	auto isSameName = [&](const QVariantMap &info) {
-		QString platFormName = getInfo(info, g_platformName);
+		QString platFormName = getInfo(info, g_channelName);
 		int channelT = getInfo(info, g_data_type, NoType);
-		return (platFormName == channelName) && (channelT == type);
+		return (platFormName == tempName) && (channelT == type);
 	};
 
 	if (auto ret = std::find_if(mChannelInfos.begin(), mChannelInfos.end(), isSameName); ret != mChannelInfos.end()) {
@@ -834,7 +854,7 @@ const QVariantMap &PLSChannelDataAPI::getChanelInfoRefBySubChannelID(const QStri
 {
 	QReadLocker locker(&mInfosLocker);
 	auto isSameID = [&](const QVariantMap &info) {
-		QString platFormName = getInfo(info, g_platformName);
+		QString platFormName = getInfo(info, g_channelName);
 		QString channelID = getInfo(info, ChannelData::g_subChannelId);
 		return (platFormName == channelName) && (channelID == subChannelID);
 	};
@@ -853,6 +873,8 @@ InfosList PLSChannelDataAPI::getChanelInfosByPlatformName(const QString &channel
 	QReadLocker locker(&mInfosLocker);
 	InfosList retLst;
 
+	auto tempName = NCB2BConvertChannelName(channelName);
+
 	for (const auto &info : mChannelInfos) {
 		bool isTypeMatch = true;
 
@@ -860,7 +882,7 @@ InfosList PLSChannelDataAPI::getChanelInfosByPlatformName(const QString &channel
 			int myType = getInfo(info, g_data_type, NoType);
 			isTypeMatch = (type == myType);
 		}
-		if (getInfo(info, g_platformName) == channelName && isTypeMatch) {
+		if (getInfo(info, g_channelName) == tempName && isTypeMatch) {
 			retLst << info;
 		}
 	}
@@ -889,7 +911,7 @@ bool PLSChannelDataAPI::isPlatformHasCountForEndView(const QString &platformName
 EndLiveList PLSChannelDataAPI::getEndLiveList(const QVariantMap &info) const
 {
 	EndLiveList ret;
-	auto platform = getInfo(info, g_platformName);
+	auto platform = getInfo(info, g_channelName);
 	auto type = getInfo(info, g_data_type, NoType);
 	if (type != ChannelType) {
 		return ret;
@@ -911,7 +933,8 @@ void PLSChannelDataAPI::updatePlatformsStates()
 {
 	QStringList platforms = getDefaultPlatforms();
 	for (auto platform : platforms) {
-		mPlatformStates.insert(platform, true);
+		auto platformName = channleNameConvertFixPlatformName(platform);
+		mPlatformStates.insert(platformName, true);
 	}
 }
 
@@ -938,6 +961,30 @@ const QMap<QString, QString> &PLSChannelDataAPI::getRTMPInfos()
 {
 	updateRtmpGpopInfos();
 	return mRtmps;
+}
+
+QStringList PLSChannelDataAPI::getUuidListOfRISTandSRT()
+{
+	QStringList uuidList;
+	QReadLocker locker(&mInfosLocker);
+	for (const auto &info : mChannelInfos) {
+		int myType = getInfo(info, g_data_type, NoType);
+		if (myType == SRTType || myType == RISTType) {
+			auto uuid = getInfo(info, g_channelUUID);
+			uuidList << uuid;
+		}
+	}
+	PLS_INFO("Channels", "get size of SRT or RIST is %d", uuidList.count());
+	return uuidList;
+}
+
+void PLSChannelDataAPI::addRISTandSRT2RtmpServer()
+{
+	QStringList uuidList = getUuidListOfRISTandSRT();
+	for (const auto &uuid : uuidList) {
+		PLS_INFO("Channels", "add SRT or RIST to new api,uuid is %s", uuid.toUtf8().constData());
+		AddOrgDataToNewApi(uuid, false);
+	}
 }
 
 std::tuple<bool, bool> PLSChannelDataAPI::isPlatformBeSurportedByCurrentVersion(const QString &platform) const
@@ -977,7 +1024,7 @@ void PLSChannelDataAPI::tryRemoveChannel(const QString &channelUUID, bool notify
 		PRE_LOG_MSG_STEP("channel has been removed ", g_removeChannelStep, WARN)
 		return;
 	}
-	auto platformName = getInfo(info, g_platformName);
+	auto platformName = getInfo(info, g_channelName);
 	bool isMultiChildren = this->isPlatformMultiChildren(platformName);
 	int myType = getInfo(info, g_data_type, NoType);
 	if (isMultiChildren && myType == ChannelType) {
@@ -1001,19 +1048,25 @@ void PLSChannelDataAPI::removeChannelInfo(const QString &channelUUID, bool notif
 	}
 
 	int channelT = getInfo(tmpInfo, g_data_type, NoType);
-	if (channelT == RTMPType && notifyServer) {
+	if (channelT >= RTMPType && notifyServer) {
 
 		RTMPDeleteToPrism(channelUUID);
 		return;
 	}
 
 	int count = removeInfo(channelUUID);
-	auto platformName = getInfo(tmpInfo, g_platformName, QString(" empty "));
-
-	QString msg = " End remmove channel " + channelUUID + " platform: " + platformName;
+	auto channelName = getInfo(tmpInfo, g_channelName, QString(" empty "));
+	auto bLeader = getInfo(tmpInfo, g_isLeader, false);
+	QString msg = " End remmove channel " + channelUUID + " channel: " + channelName;
 	PRE_LOG_MSG_STEP(msg, g_removeChannelStep, INFO)
 	if (count > 0 && notify) {
+		auto platformName = getInfo(tmpInfo, g_fixPlatformName);
+		PLSBasic::instance()->ClearService(platformName);
 		emit channelRemoved(channelUUID);
+		channelRemovedForCheckVideo(bLeader);
+		if (platformName == CHZZK) {
+			channelRemovedForChzzk();
+		}
 	}
 
 	if (channelT == ChannelType) {
@@ -1068,9 +1121,9 @@ InfosList PLSChannelDataAPI::sortAllChannels()
 			return ltype < rtype;
 		}
 
-		QString lplatform = getInfo(left, g_platformName);
+		QString lplatform = getInfo(left, g_channelName);
 
-		if (QString rplatform = getInfo(right, g_platformName); lplatform != rplatform) {
+		if (QString rplatform = getInfo(right, g_channelName); lplatform != rplatform) {
 			return isPlatformOrderLessThan(lplatform, rplatform);
 		}
 
@@ -1095,7 +1148,8 @@ InfosList PLSChannelDataAPI::sortAllChannels()
 
 void PLSChannelDataAPI::clearAll()
 {
-	emit sigAllClear();
+	PLS_INFO("Channels", "prism logout will clear all channels");
+	QWriteLocker locker(&mInfosLocker);
 	mChannelInfos.clear();
 	mImagesCache.clear();
 }
@@ -1204,9 +1258,30 @@ void PLSChannelDataAPI::reloadData()
 	if (QFile::exists(cachePath)) {
 		ChannelsMap tmp;
 		loadDataFromFile(tmp, cachePath);
+		bool bExistCHzzk = false;
 		if (!tmp.isEmpty()) {
-			mChannelInfos = std::move(tmp);
+			mChannelInfos = tmp;
 			verifyRename(mChannelInfos);
+			std::for_each(tmp.begin(), tmp.end(), [this, &bExistCHzzk](QVariantMap &info) {
+				auto type = getInfo(info, g_data_type, NoType);
+				if (type == ChannelType) {
+					auto channelName = getInfo(info, g_channelName, QString());
+					auto supportPlatforms = getDefaultPlatforms();
+					auto uuid = getInfo(info, g_channelUUID, QString());
+					if (!supportPlatforms.contains(channelName)) {
+						removeInfo(uuid);
+						return;
+					}
+					auto platform = channleNameConvertFixPlatformName(channelName);
+					setValueOfChannel(uuid, g_fixPlatformName, platform);
+					if (!bExistCHzzk && platform == CHZZK) {
+						bExistCHzzk = true;
+					}
+				}
+			});
+		}
+		if (!bExistCHzzk) {
+			channelRemovedForChzzk();
 		}
 	}
 }
@@ -1268,7 +1343,10 @@ void PLSChannelDataAPI::setRecordState(int state)
 	case RecordStopped:
 		if (!PLS_PLATFORM_API->isStopForExit()) {
 			auto isShowEnd = lastState == RecordStopping;
-			auto end = [isShowEnd]() { PLS_PLATFORM_API->showEndView(true, isShowEnd); };
+			auto end = [isShowEnd]() {
+				PLS_INFO("Channel", "singleShot setRecordState");
+				PLS_PLATFORM_API->showEndView(true, isShowEnd);
+			};
 			QTimer::singleShot(100, qApp, end);
 		}
 
@@ -1325,6 +1403,7 @@ void PLSChannelDataAPI::setBroadcastState(int event)
 	case CanBroadcastState:
 
 		if (!toGoLive()) {
+			emit startFailed();
 			this->setBroadcastState(ReadyState);
 			return;
 		}
@@ -1424,7 +1503,8 @@ void PLSChannelDataAPI::registerPlatformHandler(ChannelDataBaseHandler *handler)
 
 PlatformHandlerPtrs PLSChannelDataAPI::getPlatformHandler(const QString &platformName) const
 {
-	return mPlatformHandler.value(platformName.toUpper());
+	auto name = channleNameConvertFixPlatformName(platformName);
+	return mPlatformHandler.value(name.toUpper());
 }
 
 bool PLSChannelDataAPI::isPlatformMultiChildren(const QString &platformName) const
@@ -1439,7 +1519,30 @@ bool PLSChannelDataAPI::isPlatformMultiChildren(const QString &platformName) con
 bool PLSChannelDataAPI::isChannelMultiChildren(const QString &uuid) const
 {
 	auto info = getChannelInfo(uuid);
-	return isPlatformMultiChildren(getInfo(info, g_platformName));
+	return isPlatformMultiChildren(getInfo(info, g_channelName));
+}
+
+ChannelsMap PLSChannelDataAPI::getMatchKeysInfos(const QVariantMap &keysMap)
+{
+	ChannelsMap ret;
+	auto isMatched = [&](const QVariantMap &info) {
+		for (auto srcIte = keysMap.cbegin(); srcIte != keysMap.cend(); ++srcIte) {
+			if (srcIte.value() != getInfo(info, srcIte.key(), QVariant())) {
+				return false;
+			}
+		}
+		return true;
+	};
+
+	QReadLocker locker(&mInfosLocker);
+	for (auto ite = mChannelInfos.cbegin(); ite != mChannelInfos.cend(); ++ite) {
+		const auto &info = ite.value();
+		if (isMatched(info)) {
+			ret.insert(ite.key(), info);
+		}
+	}
+
+	return ret;
 }
 
 void PLSChannelDataAPI::stopAll()
